==> R/00_all.qmd <==
---
title: "Run all scripts"
format:
  html:
    embed-resources: true
editor: visual
---

```{r}
library("quarto")
```

1.  Load Data

```{r}
#| include: false
quarto_render(
  input = "01_load.qmd"
)
file.rename(from = "01_load.html",
            to = "../results/01_load.html")
```

2.  Clean Data

```{r}
#| include: false
quarto_render(
  input = "02_clean.qmd"
)
file.rename(from = "02_clean.html",
            to = "../results/02_clean.html")
```

3.  Augment Data

```{r}
#| include: false
quarto_render(
  input = "03_augment.qmd"
)
file.rename(from =  "03_augment.html",
            to = "../results/03_augment.html")
```

4.  Describe Data

```{r}
#| include: false
quarto_render(
  input = "04_describe.qmd"
)
file.rename(from =  "04_describe.html",
            to = "../results/04_describe.html")
```

5.  Analysis - PCA

```{r}
#| include: false
quarto_render(
  input = "05_analysis_1.qmd"
)
file.rename(from =  "05_analysis_1.html",
            to = "../results/05_analysis_1.html")
```

6.  Analysis- GSEA

    ```{r}
    #| include: false
    quarto_render(
      input = "05_analysis_2.qmd"
    )
    file.rename(from =  "05_analysis_2.html",
                to = "../results/05_analysis_2.html")
    ```

## Master Document

```{r}
file.rename(from = "00_all.html",
            to = "../results/00_all.html")
```
